requirements.txt
# Core LangChain framework
langchain>=0.1.0

# LangGraph (stateful LLM orchestration)
langgraph>=0.0.25

# Google Gemini integration
langchain-google-genai>=0.0.7
google-generativeai>=0.3.2

# For environment variable management
python-dotenv>=1.0.1

requests>=2.31.0


